fix(vnengine): address final-review minors (int-span, load-order, test gaps)
Whole-branch review (c59cd99..56e1d66) surfaced four Minor findings; this
commit clears all four:

- IRandom.Range: widen span to 64-bit so a full int-range request
  (int.MinValue..int.MaxValue) can no longer overflow to span 0 and throw a
  raw DivideByZeroException past the VnException catch. Sequences unchanged
  for all realistic ranges.
- VNRunner.LoadFromSlot: build the replacement interpreter BEFORE stopping
  the running loop, so a rebuild failure leaves the current run intact
  instead of freezing the runner.
- Test: CallStackSurvivesSaveLoad - save/restore while inside a call frame,
  proving the return address survives (callstack orientation was only
  structurally proven before).
- Test: ReadCorruptJsonReturnsNull - SaveSystem.Read returns null on
  malformed JSON.

EditMode: 157/157 green.

+++ Assets/Scripts/VNEngine/Core/Runtime/IRandom.cs
@@ -39,8 +39,11 @@ namespace VNEngine
         public int Range(int minInclusive, int maxInclusive)
         {
             if (maxInclusive < minInclusive) return minInclusive;
-            uint span = (uint)(maxInclusive - minInclusive) + 1u; // span>=1
-            return minInclusive + (int)(Next() % span);
+            // 64-bit span so the full int range (int.MinValue..int.MaxValue) can't
+            // overflow to 0 and divide-by-zero. For all normal ranges Next() % span
+            // is identical to the previous uint computation, so sequences are unchanged.
+            ulong span = (ulong)((long)maxInclusive - minInclusive) + 1UL; // span>=1
+            return (int)((long)minInclusive + (long)(Next() % span));
         }
     }
 }
